[python3 migration] Added python3 support for CT infrastructure (#4708)
Description of PR
PR cannot be merged until merging of #9277 This PR includes necessary changes for the setup of the Python3 virtual environment in the sonic-mgmt docker container.
Summary:

Moved Python modules for remove-topo from Python2 to Python3
Moved Python modules for add-topo from Python2 to Python3
Moved Python modules for deploy-mg from Python2 to Python3
Moved Python modules for start-vms from Python2 to Python3
Moved Python modules for stop-vms from Python2 to Python3
Updated bgp_fact test case for supporting Python3
How to use Python3 environment for running test cases?
Connect to the sonic-mgmt container
$ docker exec -ti sonic-mgmt bash
Activate the virtual environment
$ source /var/user/env-python3/bin/activate
Deploy t0/t1 topology
Run bgp_fact test case

+++ ansible/library/announce_routes.py
@@ -23,13 +23,13 @@ Options:
     - option-name: ptf_ip
       description: PTF container management IP address
       required: True
-      
+
     - option-name: action
       description: announce or withdraw routes
       required: False
-      
+
     - option-name: path
-      description: to figure out the path of topo_{}.yml 
+      description: to figure out the path of topo_{}.yml
       required: False
 '''
 
@@ -204,10 +204,10 @@ def generate_routes(family, podset_number, tor_number, tor_subnet_number,
                 suffix = ((podset * tor_number * max_tor_subnet_number * tor_subnet_size) + \
                           (tor * max_tor_subnet_number * tor_subnet_size) + \
                           (subnet * tor_subnet_size))
-                octet2 = (168 + (suffix / (256 ** 2)))
-                octet1 = (192 + (octet2 / 256))
+                octet2 = (168 + int(suffix / (256 ** 2)))
+                octet1 = (192 + int(octet2 / 256))
                 octet2 = (octet2 % 256)
-                octet3 = ((suffix / 256) % 256)
+                octet3 = (int(suffix / 256) % 256)
                 octet4 = (suffix % 256)
                 prefixlen_v4 = (32 - int(math.log(tor_subnet_size, 2)))
